Upload Initial Version of Project I
Uploaded the initial version of Project I, the 32-bit MIPS Simulator with GUI Dashboard. However, there are some issues that require attention and fixing. This commit serves as a starting point and will be followed by subsequent updates to address the identified problems and improve the overall functionality and clarity of the project.